pyproject.toml
[project]
name = "naver-smartstore-faq-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.6",
    "openai>=1.57.0",
    "pymilvus>=2.5.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.40.2",
]
